homogeneize geospatial training code
