--- conf/conf_geospatial/dataset/base.yaml
@@ -1,5 +1,0 @@
-standardize: true
-squash: false
-randomize_x0: false
-virtual_node: false
-add_positions: false
