+++ conf/conf_geospatial/dataset/cc-virtual.yaml
@@ -1,0 +1,2 @@
+squash_to_graph: false
+virtual_node: true
