+++ conf/conf_geospatial/dataset/cc.yaml
@@ -1,0 +1,2 @@
+squash_to_graph: false
+virtual_node: false
