--- conf/conf_geospatial/dataset/randomize-x0.yaml
@@ -1,5 +1,0 @@
-standardize: true
-squash: false
-randomize_x0: true
-virtual_node: false
-add_positions: false
