--- conf/conf_geospatial/dataset/virtual-node.yaml
@@ -1,5 +1,0 @@
-standardize: true
-squash: false
-randomize_x0: false
-virtual_node: true
-add_positions: false
