+++ conf/conf_geospatial/experiment/egnn-equivariant.yaml
@@ -1,0 +1,7 @@
+#@package _global_
+defaults:
+  - override /dataset: graph
+
+model:
+  pos_update: true
+  hausdorff: false
