+++ conf/conf_geospatial/experiment/equivariant-virtual.yaml
@@ -1,0 +1,7 @@
+#@package _global_
+defaults:
+  - override /dataset: cc-virtual
+
+model:
+  pos_update: true
+  hausdorff: true
