+++ conf/conf_geospatial/experiment/equivariant.yaml
@@ -1,0 +1,7 @@
+#@package _global_
+defaults:
+  - override /dataset: cc
+
+model:
+  pos_update: true
+  hausdorff: true
