+++ conf/conf_geospatial/model/large.yaml
@@ -1,0 +1,2 @@
+num_hidden: 32
+num_layers: 4
