+++ etnn/utils.py
@@ -1,6 +1,7 @@
 from typing import Optional
 
 import numpy as np
+from numba import numba
 import torch
 from torch import Tensor
