+++ main_geospatial.py
@@ -1,20 +1,21 @@
+from collections import defaultdict
 import copy
-import json
 import logging
 import os
-import random
 import time
+from functools import partial
 
 import hydra
+import numpy as np
 import torch
 from omegaconf import DictConfig, OmegaConf
 from torch_geometric.loader import DataLoader
+from torch_geometric.transforms import Compose
 from tqdm import tqdm
 
 import utils
 import wandb
-from etnn.geospatial.pm25cc import PM25CC
-
+from etnn.geospatial import pm25cc, transforms
 
 # torch.set_float32_matmul_precision("high")  # Use high precision for matmul
 os.environ["WANDB__SERVICE_WAIT"] = "600"
@@ -33,19 +34,31 @@ def main(cfg: DictConfig):
     # create a unique hash for the dataset based on the configuration
     hash = utils.args_to_hash(OmegaConf.to_container(cfg.dataset, resolve=True))
 
-    dataset = PM25CC(
-        f"data/qm9cc_{hash}",
-        lifters=list(cfg.dataset.lifters),
-        neighbor_types=list(cfg.dataset.neighbor_types),
-        connectivity=cfg.dataset.connectivity,
-        # cfg.lifter.dim,
-        # list(cfg.lifter.initial_features),
-        # merge_neighbors=cfg.model.merge_neighbors,
-        supercell=cfg.dataset.supercell,
-        force_reload=False,
-        transform=prepare_targets_transform,
-    )
+    pre_transform = []
+    if cfg.dataset.standardize:
+        pre_transform.append(transforms.standardize_cc)
+    if cfg.dataset.randomize_x0:
+        pre_transform.append(partial(transforms.randomize, keys=["x_0"]))
+    if cfg.dataset.virtual_node:
+        pre_transform.append(transforms.add_virtual_node)
+    if cfg.dataset.squash_to_graph:
+        pre_transform.append(transforms.squash_cc)
+    if cfg.dataset.add_positions:
+        pre_transform.append(transforms.add_pos_to_cc)
+    pre_transform = Compose(pre_transform)
 
+    # mask a fraction of node during prediction, needed for node-level tasks
+    masking_transform = transforms.create_mask(seed=cfg.seed, rate=cfg.mask_rate)
+
+    dataset = pm25cc.PM25CC(
+        f"data/geospatialcc_{hash}",
+        pre_transform=pre_transform,
+        force_reload=cfg.force_reload,
+        transform=masking_transform,
+    )
+    logger.info(
+        f"Created GeospatialCC dataset generated and stored in '{dataset.root}'."
+    )
     # ==== Get model =====
     model = utils.get_model(cfg, dataset)
 
@@ -53,43 +66,11 @@ def main(cfg: DictConfig):
     logger.info(f"Number of parameters: {num_params:}")
     logger.info(model)
 
-    # Get train/test splits using the original egnn splits for reference
-    with open("data/input/egnn_splits.json", "r") as io:
-        egnn_splits = json.load(io)
-    if cfg.train_test_splits == "egnn":
-        split_indices = egnn_splits
-
-    elif cfg.train_test_splits == "random":
-        num_samples = len(dataset)
-        indices = list(range(num_samples))
-        random.shuffle(indices)
-        train_end_idx = len(egnn_splits["train"])
-        val_end_idx = train_end_idx + len(egnn_splits["valid"])
-        test_end_idx = val_end_idx + len(egnn_splits["test"])
-        split_indices = {
-            "train": indices[:train_end_idx],
-            "valid": indices[train_end_idx:val_end_idx],
-            "test": indices[val_end_idx:test_end_idx],
-        }
-    else:
-        raise ValueError(f"Unknown split type: {cfg.train_test_splits}")
-
     # Get dataloaders
-    loaders = {
-        key: DataLoader(
-            dataset[indices],
-            batch_size=cfg.training.batch_size,
-            shuffle=True,
-        )
-        for key, indices in split_indices.items()
-    }
-
-    # Precompute average deviation of target in loader
-    mean, mad = utils.calc_mean_mad(loaders["train"])
-    mean, mad = mean.to(device), mad.to(device)
+    loader = DataLoader(dataset, batch_size=1, shuffle=False)
 
     # ==== Get optimization objects =====
-    crit = torch.nn.L1Loss(reduction="mean")
+    crit = torch.nn.MSELoss(reduction="mean")
     opt_kwargs = dict(lr=cfg.training.lr, weight_decay=cfg.training.weight_decay)
     opt = torch.optim.Adam(model.parameters(), **opt_kwargs)
     T_max = cfg.training.epochs // cfg.training.num_lr_cycles
@@ -133,40 +114,63 @@ def main(cfg: DictConfig):
 
     # === Training loop ===
     for epoch in tqdm(range(start_epoch, cfg.training.epochs)):
-        epoch_start_time, epoch_mae_train, epoch_mae_val = time.time(), 0, 0
+        epoch_start_time = time.time()
+
+        epoch_metrics = defaultdict(list)
 
         model.train()
-        for _, batch in enumerate(loaders["train"]):
+        for _, batch in enumerate(loader):
             opt.zero_grad()
             batch = batch.to(device)
 
-            pred = model(batch)
-            loss = crit(pred, (batch.y - mean) / mad)
-            mae = crit(pred * mad + mean, batch.y)
-            loss.backward()
-
+            # == training step ==
+            opt.zero_grad()
+            outputs = model(batch)
+            training_mask = batch.training_mask  # 1-0 mask of nodes used for training
+            pred_train = outputs["0"].squeeze()[training_mask]
+            target_train = batch.y.squeeze()[training_mask]
+            loss_terms = crit(pred_train, target_train)
+            train_loss = loss_terms.mean()
+            train_loss.backward()  # backpropagate
             if cfg.training.clip_gradients:
-                torch.nn.utils.clip_grad_norm_(
+                torch.nn.utils.clip_grad_value_(
                     model.parameters(), cfg.training.clip_amount
                 )
-
             opt.step()
-            epoch_mae_train += mae.item()
+
+            if torch.cuda.is_available():
+                torch.cuda.empty_cache()
+
+            # == eval step ===
+            model.eval()
+            if model.dropout > 0:
+                with torch.no_grad():
+                    pred_eval = model(batch)["0"].squeeze()
+            else:
+                pred_eval = outputs["0"].squeeze().detach()
+            pred_test = pred_eval[batch.test_mask]
+            pred_val = pred_eval[batch.validation_mask]
+            target_test = batch.y.squeeze()[batch.test_mask]
+            target_val = batch.y.squeeze()[batch.validation_mask]
+
+            val_r2 = 1 - (pred_val - target_val).var() / target_val.var()
+            test_r2 = 1 - (pred_test - target_test).var() / target_test.var()
+            test_mse = (pred_test - target_test).pow(2).mean()
+            val_mse = (pred_val - target_val).pow(2).mean()
+
+            epoch_metrics["train_loss"].append(train_loss.item())
+            epoch_metrics["val_r2"].append(val_r2.item())
+            epoch_metrics["val_mse"].append(val_mse.item())
+            epoch_metrics["test_r2"].append(test_r2.item())
+            epoch_metrics["test_mse"].append(test_mse.item())
+            epoch_metrics["lr"].append(opt.param_groups[0]["lr"])
 
         sched.step()
-        model.eval()
-        for _, batch in enumerate(loaders["valid"]):
-            batch = batch.to(device)
-            pred = model(batch)
-            mae = crit(pred * mad + mean, batch.y)
 
-            epoch_mae_val += mae.item()
+        epoch_loss_val = np.mean(epoch_metrics["val_mse"])
 
-        epoch_mae_train /= len(loaders["train"])
-        epoch_mae_val /= len(loaders["valid"])
-
-        if epoch_mae_val < best_loss:
-            best_loss = epoch_mae_val
+        if epoch_loss_val < best_loss:
+            best_loss = epoch_loss_val
             best_model = copy.deepcopy(model)
 
         # Save checkpoint
@@ -184,49 +188,17 @@ def main(cfg: DictConfig):
         epoch_end_time = time.time()
         epoch_duration = epoch_end_time - epoch_start_time
 
+        epoch_metrics = {k: np.nanmean(v) for k, v in epoch_metrics.items()}
+
         wandb.log(
             {
-                "Train MAE": epoch_mae_train,
-                "Validation MAE": epoch_mae_val,
                 "Epoch Duration": epoch_duration,
                 "Learning Rate": sched.get_last_lr()[0],
+                "Epoch": epoch + 1,
+                **epoch_metrics,
             },
             step=epoch,
         )
-
-        # Compute and log test error every test_interval epochs
-        if (epoch + 1) % cfg.training.test_interval == 0:
-            test_mae = 0
-            best_model.eval()
-            for _, batch in enumerate(loaders["test"]):
-                batch = batch.to(device)
-                pred = best_model(batch)
-                mae = crit(pred * mad + mean, batch.y)
-                test_mae += mae.item()
-
-            test_mae /= len(loaders["test"])
-            print(f"Epoch {epoch + 1} Test MAE: {test_mae}")
-
-            wandb.log(
-                {
-                    "Interval Test MAE": test_mae,
-                    "Epoch": epoch + 1,
-                },
-                step=epoch,
-            )
-
-    test_mae = 0
-    best_model.eval()
-    for _, batch in enumerate(loaders["test"]):
-        batch = batch.to(device)
-        pred = best_model(batch)
-        mae = crit(pred * mad + mean, batch.y)
-        test_mae += mae.item()
-
-    test_mae /= len(loaders["test"])
-    print(f"Test MAE: {test_mae}")
-
-    wandb.log({"Test MAE": test_mae})
 
 
 if __name__ == "__main__":
